Extract new class "BoardDrawer" from BoardController and make it a field in BoardController
