+++ src/main/java/data/board/BoardController.java
@@ -6,71 +6,33 @@ import data.round.roundState.RoundState;
 public class BoardController {
     final private BoardData gameBoardData;
     final private GameRoundStateChecker gameRoundStateChecker;
-    private int fieldCounter;
+    final private BoardDrawer boardDrawer;
 
     public BoardController(BoardData gameBoardData, int howManySymbolsInUnbrokenLineToWin) {
         this.gameBoardData = gameBoardData;
         gameRoundStateChecker = new GameRoundStateChecker(howManySymbolsInUnbrokenLineToWin);
+        boardDrawer = new BoardDrawer();
     }
 
     @Override
     public String toString() {
-        StringBuilder tableDisplayInfoBuilder = new StringBuilder();
-        tableDisplayInfoBuilder.append(getDisplayTopLineOfBoard()).append("\n");
+        return calculateStringDisplayingGameBoard();
+    }
 
-        fieldCounter = 1;
+    private String calculateStringDisplayingGameBoard() {
+        StringBuilder tableDisplayInfoBuilder = new StringBuilder();
+        tableDisplayInfoBuilder.append(boardDrawer.getDisplayTopLineOfBoard(gameBoardData)).append("\n");
+
+        int curFirstLeftColumnNumber = 1;
         for (int levelCountingFromTop = 0; levelCountingFromTop < gameBoardData.height; ++levelCountingFromTop){
             tableDisplayInfoBuilder
-                    .append(getDisplayLine_withEmptySpacesInsideField()).append("\n")
-                    .append(getDisplayLine_withSymbolOrNumberInsideField()).append("\n")
-                    .append(getDisplayLine_verticalSeparatorBetweenFields()).append("\n");
+                    .append(boardDrawer.getDisplayLine_withEmptySpacesInsideField(gameBoardData)).append("\n")
+                    .append(boardDrawer.getDisplayLine_withSymbolOrNumberInsideField(gameBoardData, curFirstLeftColumnNumber)).append("\n")
+                    .append(boardDrawer.getDisplayLine_verticalSeparatorBetweenFields(gameBoardData)).append("\n");
+            curFirstLeftColumnNumber += gameBoardData.width;
         }
         return tableDisplayInfoBuilder.toString();
     }
-
-    private String getDisplayLine_withSymbolOrNumberInsideField() {
-        StringBuilder displayLineWithEmptySpaces = new StringBuilder();
-        for (int i = 0; i < gameBoardData.width; ++i) {
-            Symbol symbolOnField = gameBoardData.getSymbolFromField(fieldCounter);
-
-            displayLineWithEmptySpaces.append("|");
-
-            if (symbolOnField != null || String.valueOf(fieldCounter).length() < 3) {
-                displayLineWithEmptySpaces.append(" ");
-            }
-
-            displayLineWithEmptySpaces.append(symbolOnField == null ? fieldCounter : symbolOnField);
-
-            if (symbolOnField != null || String.valueOf(fieldCounter).length() < 2) {
-                displayLineWithEmptySpaces.append(" ");
-            }
-
-            ++fieldCounter;
-        }
-        return displayLineWithEmptySpaces.append("|").toString();
-    }
-
-    private String getDisplayTopLineOfBoard() {
-        return new String(new char[gameBoardData.width*4+1]).replace("\0", "_");
-    }
-
-    private String getDisplayLine_verticalSeparatorBetweenFields() {
-        return getHorizontalLineOfBoardDisplayWithHorizontalSeparators("___");
-    }
-
-    private String getDisplayLine_withEmptySpacesInsideField() {
-        return getHorizontalLineOfBoardDisplayWithHorizontalSeparators("   ");
-    }
-
-    private String getHorizontalLineOfBoardDisplayWithHorizontalSeparators(String spaceBetweenSeparators) {
-        StringBuilder displayLineWithEmptySpaces = new StringBuilder();
-        for (int i = 0; i < gameBoardData.width; ++i) {
-            displayLineWithEmptySpaces.append("|")
-                    .append(spaceBetweenSeparators);
-        }
-        return displayLineWithEmptySpaces.append("|").toString();
-    }
-
 
     public boolean tryMarkFieldAndChangeWinnerStateIfNeeded(int fieldNumber, Symbol symbol) {
         if (gameBoardData.isFieldOccupied(fieldNumber) || fieldNumber < 1 || fieldNumber > gameBoardData.width*gameBoardData.height)
