--- src/test/java/data/round/roundState/RoundStateDrawTest.java
@@ -1,7 +1,0 @@
-package data.round.roundState;
-
-import static org.testng.Assert.*;
-
-public class RoundStateDrawTest {
-
-}
